Document and implement common AngularJS patterns
Implemented a component with a controller and template and states.

This integrates the accounts-password and useraccounts:unstyled with the
boilerplate.  The useraccounts is a blaze template and blazeTemplate
directive was used to incorporate it into an AngularJS app.

This adds the notion of an abstract state representing authenticated and
unauthenticated.  In addition, the atForm is removed from home and put
into the sign-in state.  The currentUser is shown on the home state.

The onSubmitHook is implemented to better integrate the AccountsTemplates
workflow with the boilerplate.

This sets up JSDoc memberof tags to represent the documentation as a tree
better.

JSDoc should be present on the top of the JS file.